# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "API ConfApp",
        "contact": {
            "name": "Company",
            "url": "https://example.ru/",
            "email": "example@mail.ru"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:22952",
    "basePath": "/api",
    "paths": {
        "/v1/config": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "получить конфиг по названию сервиса и номеру версии конфига",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название сервиса удаляемого конфига",
                        "name": "service",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Номер версии",
                        "name": "v",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "обновить одну версию конфига по названию сервиса и номеру версии",
                "parameters": [
                    {
                        "description": "Заменить кавычки на двойные - {'service': 'managed-k8s', 'v': 224, 'data': [{'key1': 'value1-1'}]}",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "type": "primitive"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create"
                ],
                "summary": "создать конфиг для нового сервиса",
                "parameters": [
                    {
                        "description": "Заменить кавычки на двойные - {'service': 'managed-k8s', 'data': [{'key1': 'value1'}, {'key2': 'value2'}]}",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "type": "primitive"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "delete"
                ],
                "summary": "удалить версию конфига по названию сервиса(помечает удалённым, удаляет через 90 дней)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название сервиса удаляемого конфига",
                        "name": "service",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Номер версии",
                        "name": "v",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    }
                }
            }
        },
        "/v1/config/any": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "обновить и/или добавить новые поля во все версии конфига сервиса",
                "parameters": [
                    {
                        "description": "Заменить кавычки на двойные - {'service': 'managed-k8s', 'data': [{'key1': 'changed_value'}]}",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "type": "primitive"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    }
                }
            }
        },
        "/v1/config/update": {
            "post": {
                "description": "Создаёт копию последней версии с новыми и/или измененными параметрами, под новым номером версии",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create"
                ],
                "summary": "обновить последний конфиг",
                "parameters": [
                    {
                        "description": "Заменить кавычки на двойные - {'service': 'managed-k8s', 'data': [{'key3': 'value3'}]}",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "type": "primitive"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    }
                }
            }
        },
        "/v1/service": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "delete"
                ],
                "summary": "удалить сервис по названию(помечает удалёнными все версии, удаляет через 90 дней)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя удаляемого сервиса",
                        "name": "service",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.HTTPStatus"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.HTTPStatus": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}